Update imports to use aliases
